separated snapshot tests into new file
